sm8250-common: livedisplay: switch to opt-in for AdaptiveBacklight, AntiFlicker and SunlightEnhancement interfaces
The previous implementation with 1 vintf fragment containing AdaptiveBacklight,
PictureAdjustment and SunlightEnhancement would cause the following logcat spam on
pstar because AdaptiveBacklight is not supported.

W HidlServiceManagement: Waited one second for vendor.lineage.livedisplay@2.0::IAdaptiveBacklight/default
I hwservicemanager: Since vendor.lineage.livedisplay@2.0::IAdaptiveBacklight/default is not registered, trying to start it as a lazy HAL.

While also causing AntiFlicker to fail to start, as it isn't defined in the vintf fragment.

I hwservicemanager: getTransport: Cannot find entry vendor.lineage.livedisplay@2.1::IAntiFlicker/default in either framework or device VINTF manifest.
E HidlServiceManagement: Service vendor.lineage.livedisplay@2.1::IAntiFlicker/default must be in VINTF manifest in order to register/get.

Both of which will lead to the phone to hang on "Phone is starting", not loading the launcher,
and eventually soft rebooting.

As pstar and nio support a slightly different set of interfaces, making those interfaces
opt-in solves the errors mentioned above.
pstar supports AntiFlicker exclusively, while nio supports AdaptiveBacklight exclusively.
SunlightEnhancement, which is supported by both nio and pstar, is opt-in for now
as the LED on pstar bleeds into the light sensor, triggering the high brightness mode
when the LED is active during charging or due to notifications.

Change-Id: I18545c62a7c698c28c42d377d98c6c8bdd3b910d